batch_processing/requirements.txt
absl-py==0.9.0
addignore==1.2.7
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
black==19.10b0
bokeh==1.4.0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
deepspeech==0.8.0
delegator.py @ git+https://github.com/amitt001/delegator.py.git@194aa92543fbdbfbae0bcc24ca217819a7805da2
flask==1.1.2
gast==0.2.2
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
isort==4.3.21
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pathspec==0.8.0
pexpect==4.8.0
phonemizer==2.2
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.3.1
regex==2020.6.7
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
soundfile==0.10.3.post1
tensorboard==2.1.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.1.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
toml==0.10.1
tqdm==4.46.1
tts==0.0.2+f320992
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
